Fixed Shields not having Subtypes, and Weapons gaining the Masterwork property when enhanced.
Shields not having subtypes was causing custom shield creation to fail: in addition, shields have the armor type and shield subtype in FG, so that is what the types are set to.

In FG, weapons with +Enhancement Bonuses and the Masterwork property gain the bonuses of both, which is not how it works.